--- packages/ultrahonk/src/polynomials/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod polynomial_types;
